requirements.txt
python-dotenv
langchain
openai==1.3.3
langchain-openai
pytest
